+++ packages/di/src/injection/dependency.ts
@@ -1,5 +1,4 @@
 import { Key } from "./key";
-import { SelfAnnotation, SkipSelfAnnotation } from "./metadata";
 
 /**
  * `Dependency` is used by the framework to extend DI.
@@ -9,12 +8,12 @@ export class Dependency {
     constructor(
         public key: Key,
         public optional: boolean,
-        public visibility: SelfAnnotation | SkipSelfAnnotation | null,
+        public visibility: 'self' | 'skip-self' | 'default',
         public skip: boolean
     ) {
     }
 
     static fromKey(key: Key): Dependency {
-        return new Dependency(key, false, null, false);
+        return new Dependency(key, false, 'default', false);
     }
 }