+++ packages/di/src/injection/forward-ref.ts
@@ -1,53 +1,8 @@
-import { stringify } from './stringify';
+import { isConstructor } from "@alterior/functions";
 
-/**
- * An interface that a function passed into {@link forwardRef} has to implement.
- *
- * ### Example
- *
- * {@example core/di/ts/forward_ref/forward_ref_spec.ts region='forward_ref_fn'}
- * @experimental
- */
-export interface ForwardRefFn<T = any> {
-  (): T;
-}
-
-/**
- * Allows to refer to references which are not yet defined.
- *
- * For instance, `forwardRef` is used when the `token` which we need to refer to for the purposes of
- * DI is declared,
- * but not yet defined. It is also used when the `token` which we use when creating a query is not
- * yet defined.
- *
- * ### Example
- * {@example core/di/ts/forward_ref/forward_ref_spec.ts region='forward_ref'}
- * @experimental
- */
-export function forwardRef<T>(forwardRefFn: ForwardRefFn): T {
-  (<any>forwardRefFn).__forward_ref__ = forwardRef;
-  (<any>forwardRefFn).toString = function() {
-    return stringify(this());
-  };
-  return <T>(<any>forwardRefFn);
-}
-
-/**
- * Lazily retrieves the reference value from a forwardRef.
- *
- * Acts as the identity function when given a non-forward-ref value.
- *
- * ### Example ([live demo](http://plnkr.co/edit/GU72mJrk1fiodChcmiDR?p=preview))
- *
- * {@example core/di/ts/forward_ref/forward_ref_spec.ts region='resolve_forward_ref'}
- *
- * See: {@link forwardRef}
- * @experimental
- */
-export function resolveForwardRef(type: any): any {
-  if (typeof type === 'function' && type.hasOwnProperty('__forward_ref__') && type.__forward_ref__ === forwardRef) {
-    return (type as ForwardRefFn)();
-  } else {
-    return type;
-  }
+export function resolveForwardRef<T>(ref: T | (() => T)): T {
+    if (typeof ref === 'function' && !isConstructor(ref))
+        return (ref as (() => T))();
+    
+    return ref as T;
 }