+++ packages/di/src/injection/inject.ts
@@ -10,9 +10,13 @@ import { THROW_IF_NOT_FOUND } from "./throw-if-not-found";
  * 
  * @param token 
  * @param options 
- * @returns 
+ * @returns The resolved value of the given injection token. If the `optional` option is true and there is no provider
+ *          for the given token, `null` is returned.
  */
-export function inject<T = unknown>(token: Type<T> | InjectionToken<T>, options?: InjectOptions): T {
+export function inject<T = unknown>(token: Type<T> | InjectionToken<T>): T;
+export function inject<T = unknown>(token: Type<T> | InjectionToken<T>, options: InjectOptions & { optional: true }): T | null;
+export function inject<T = unknown>(token: Type<T> | InjectionToken<T>, options: InjectOptions & { optional?: false | undefined }): T;
+export function inject<T = unknown>(token: Type<T> | InjectionToken<T>, options?: InjectOptions): T | null {
     let injector = injectionContext().injector;
   
     if (options?.skipSelf ?? false)
@@ -20,7 +24,7 @@ export function inject<T = unknown>(token: Type<T> | InjectionToken<T>, options?
   
     return injector.get(
       token, 
-      options?.optional === true ? undefined : <any>THROW_IF_NOT_FOUND,
+      options?.optional === true ? null : <any>THROW_IF_NOT_FOUND,
       { self: options?.self ?? false }
     );
   }