+++ packages/di/src/injection/metadata/index.ts
@@ -1,2 +1,0 @@
-export * from './annotations';
-export * from './decorators';